+++ src/Gerente.java
@@ -3,7 +3,7 @@ public  class Gerente extends Funcionario {
 
 
     //metodo construtor
-    public Gerente(String nome, String cpf, float saldo) {
+    public Gerente(String nome, String cpf, double saldo) {
         super(nome, cpf, saldo);
     }
